1. INTERNAL_OPAMP_CONFIG is added to the code 2. SPEEDCNTR_ITERM, SPEEDCNTR_OUTMIN values are modified 3. For Speed Control, instead of mcappData.movingAvgFilterSpeed.avg, mcappData.calculateSpeed.speedValue is used to improve transient response
git-svn-id: https://chn-vm-apps.microchip.com/mchprepo/MCU16/applications/motor_control/MC_Projects/AN957/branches/AN957_dsPIC33CK256MP508_EXT_INT_OPAMP_LVMC@97568 f217b49b-f41f-0410-bcdc-e4b5239f133b

+++ hal/board_service.c
@@ -293,7 +293,7 @@ void HAL_MC1PWMSetDutyCyclesIdentical(int16_t dutyCycle)
   
 void HAL_MC1MotorInputsRead(MCAPP_MEASURE_T *pMotorInputs)
 {   
-    pMotorInputs->measureCurrent.Ibus = (ADCBUF_IBUS);
+    pMotorInputs->measureCurrent.Ibus = (int16_t)(ADCBUF_IBUS);
     pMotorInputs->measurePot = (int16_t)(ADCBUF_POT >>1);
     pMotorInputs->measureVdc.value = (int16_t)(ADCBUF_VBUS >>1);
 }
